+++ src/teehr/loading/nwm30/nwm_grid_data.py
@@ -1,35 +1,35 @@
 from typing import Union, Iterable, Optional
 from datetime import datetime
 
-from teehr.loading.nwm_common.point_utils import fetch_and_format_nwm_points
-
+from teehr.loading.nwm_common.grid_utils import (
+    fetch_and_format_nwm_grids
+)
 from teehr.loading.nwm_common.utils_nwm import (
-    build_remote_nwm_filelist,
     build_zarr_references,
+    build_remote_nwm_filelist,
 )
-from teehr.models.loading.nwm22_point import PointConfigurationModel
-from teehr.loading.nwm22.const_nwm import (
-    NWM22_UNIT_LOOKUP,
-    NWM22_ANALYSIS_CONFIG
-)
+from teehr.loading.nwm22.const_nwm import NWM22_UNIT_LOOKUP
+
+from teehr.models.loading.nwm30_grid import GridConfigurationModel
+from teehr.loading.nwm30.const_nwm import NWM30_ANALYSIS_CONFIG
 
 
-def nwm_to_parquet(
+def nwm_grids_to_parquet(
     configuration: str,
     output_type: str,
     variable_name: str,
     start_date: Union[str, datetime],
     ingest_days: int,
-    location_ids: Iterable[int],
+    zonal_weights_filepath: str,
     json_dir: str,
     output_parquet_dir: str,
     t_minus_hours: Optional[Iterable[int]] = None,
-    process_by_z_hour: Optional[bool] = True,
-    stepsize: Optional[int] = 100,
     ignore_missing_file: Optional[bool] = True,
     overwrite_output: Optional[bool] = False,
 ):
-    """Fetches NWM point data, formats to tabular, and saves to parquet
+    """
+    Fetches NWM gridded data, calculates zonal statistics (mean) of selected
+    variable for given zones, converts and saves to TEEHR tabular format
 
     Parameters
     ----------
@@ -47,8 +47,9 @@ def nwm_to_parquet(
         Str formats can include YYYY-MM-DD or MM/DD/YYYY
     ingest_days : int
         Number of days to ingest data after start date
-    location_ids : Iterable[int]
-        Array specifying NWM IDs of interest
+    zonal_weights_filepath: str
+        Path to the array containing fraction of pixel overlap
+        for each zone
     json_dir : str
         Directory path for saving json reference files
     output_parquet_dir : str
@@ -56,17 +57,6 @@ def nwm_to_parquet(
     t_minus_hours: Optional[Iterable[int]]
         Specifies the look-back hours to include if an assimilation
         configuration is specified.
-    process_by_z_hour: bool
-        A boolean flag that determines the method of grouping files
-        for processing. The default is True, which groups by day and z_hour.
-        False groups files sequentially into chunks, whose size is determined
-        by stepsize. This allows users to process more data potentially more
-        efficiently, but runs to risk of splitting up forecasts into separate
-        output files.
-    stepsize: int
-        The number of json files to process at one time. Used if
-        process_by_z_hour is set to False. Default value is 100. Larger values
-        can result in greater efficiency but require more memory
     ignore_missing_file: bool
         Flag specifying whether or not to fail if a missing NWM file is encountered
         True = skip and continue
@@ -76,7 +66,7 @@ def nwm_to_parquet(
         exist.  True = overwrite; False = fail
 
     The NWM configuration variables, including configuration, output_type, and
-    variable_name are stored as pydantic models in point_config_models.py
+    variable_name are stored as a pydantic model in grid_config_models.py
 
     Forecast and assimilation data is grouped and saved one file per reference
     time, using the file name convention "YYYYMMDDTHHZ".  The tabular output
@@ -92,7 +82,8 @@ def nwm_to_parquet(
             output_type: variable_name,
         },
     }
-    cm = PointConfigurationModel.parse_obj(vars)
+
+    cm = GridConfigurationModel.parse_obj(vars)
     configuration = cm.configuration.name
     forecast_obj = getattr(cm, configuration)
     output_type = forecast_obj.output_type.name
@@ -103,7 +94,7 @@ def nwm_to_parquet(
         output_type,
         start_date,
         ingest_days,
-        NWM22_ANALYSIS_CONFIG,
+        NWM30_ANALYSIS_CONFIG,
         t_minus_hours,
         ignore_missing_file,
     )
@@ -112,14 +103,12 @@ def nwm_to_parquet(
                                        json_dir,
                                        ignore_missing_file)
 
-    fetch_and_format_nwm_points(
+    fetch_and_format_nwm_grids(
         json_paths,
-        location_ids,
         configuration,
         variable_name,
         output_parquet_dir,
-        process_by_z_hour,
-        stepsize,
+        zonal_weights_filepath,
         ignore_missing_file,
         NWM22_UNIT_LOOKUP,
         overwrite_output,
@@ -127,46 +116,20 @@ def nwm_to_parquet(
 
 
 if __name__ == "__main__":
-    configuration = (
-        "short_range"  # analysis_assim_extend, short_range
-    )
-    output_type = "channel_rt"
-    variable_name = "streamflow"
-    start_date = "2023-08-24"
-    ingest_days = 1
-    location_ids = [
-        7086109,
-        7040481,
-        7053819,
-        7111205,
-        7110249,
-        14299781,
-        14251875,
-        14267476,
-        7152082,
-        14828145,
-    ]
-    # location_ids = np.load(
-    #     "/mnt/sf_shared/data/ciroh/temp_location_ids.npy"
-    # )  # all 2.7 million
-    json_dir = "/mnt/sf_shared/data/ciroh/jsons"
-    output_parquet_dir = "/mnt/sf_shared/data/ciroh/parquet"
-
-    process_by_z_hour = True
-    stepsize = 100
+    # Local testing
+    single_filepath = "/mnt/data/ciroh/nwm.20201218_forcing_short_range_nwm.t00z.short_range.forcing.f001.conus.nc"  # noqa
+    weights_parquet = "/mnt/data/ciroh/wbdhuc10_weights.parquet"
     ignore_missing_file = False
 
-    nwm_to_parquet(
-        configuration,
-        output_type,
-        variable_name,
-        start_date,
-        ingest_days,
-        location_ids,
-        json_dir,
-        output_parquet_dir,
-        t_minus_hours=[0, 1, 2],
-        process_by_z_hour=process_by_z_hour,
-        stepsize=stepsize,
-        ignore_missing_file=ignore_missing_file
+    nwm_grids_to_parquet(
+        "forcing_analysis_assim",
+        "forcing",
+        "RAINRATE",
+        "2020-12-18",
+        1,
+        weights_parquet,
+        "/home/sam/forcing_jsons",
+        "/home/sam/forcing_parquet",
+        [0],
+        ignore_missing_file,
     )