+++ src/teehr/loading/nwm_common/utils_nwm.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from typing import Union, Optional, Iterable, List
+from typing import Union, Optional, Iterable, List, Dict
 from datetime import datetime
 
 import dask
@@ -10,11 +10,47 @@ import pandas as pd
 import numpy as np
 import xarray as xr
 import geopandas as gpd
+import pyarrow as pa
+import pyarrow.parquet as pq
 
 from teehr.loading.nwm22.const_nwm import (
-    NWM22_ANALYSIS_CONFIG,
     NWM_BUCKET,
 )
+
+
+def write_parquet_file(
+    filepath: Path,
+    overwrite_output: bool,
+    data: Union[pa.Table, pd.DataFrame]
+):
+    """Writes output timeseries parquet file with logic controlling
+    whether or not to overwrite an existing file.
+
+    Parameters
+    ----------
+    filepath : Path
+        Path to the output parquet file
+    overwrite : bool
+        Flag controlling overwrite behavior
+    data : Union[pa.Table, pd.DataFrame]
+        The output data as either a dataframe or pyarrow table
+    """
+    if not filepath.is_file():
+        if isinstance(data, pa.Table):
+            pq.write_table(data, filepath)
+        else:
+            data.to_parquet(filepath)
+    elif filepath.is_file() and overwrite_output:
+        print(f"Overwriting {filepath.name}")
+        if isinstance(data, pa.Table):
+            pq.write_table(data, filepath)
+        else:
+            data.to_parquet(filepath)
+    elif filepath.is_file() and not overwrite_output:
+        print(
+            f"{filepath.name} already exists and overwrite_output=False;"
+            " skipping"
+        )
 
 
 def load_gdf(filepath: Union[str, Path], **kwargs: str) -> gpd.GeoDataFrame:
@@ -304,6 +340,7 @@ def build_remote_nwm_filelist(
     output_type: str,
     start_dt: Union[str, datetime],
     ingest_days: int,
+    analysis_config_dict: Dict,
     t_minus_hours: Optional[Iterable[int]],
     ignore_missing_file: Optional[bool],
 ) -> List[str]:
@@ -339,12 +376,12 @@ def build_remote_nwm_filelist(
     dates = pd.date_range(start=start_dt, periods=ingest_days, freq="1d")
 
     if "assim" in configuration:
-        cycle_z_hours = NWM22_ANALYSIS_CONFIG[configuration]["cycle_z_hours"]
-        domain = NWM22_ANALYSIS_CONFIG[configuration]["domain"]
-        configuration_name_in_filepath = NWM22_ANALYSIS_CONFIG[configuration][
+        cycle_z_hours = analysis_config_dict[configuration]["cycle_z_hours"]
+        domain = analysis_config_dict[configuration]["domain"]
+        configuration_name_in_filepath = analysis_config_dict[configuration][
             "configuration_name_in_filepath"
         ]
-        max_lookback = NWM22_ANALYSIS_CONFIG[configuration]["num_lookback_hrs"]
+        max_lookback = analysis_config_dict[configuration]["num_lookback_hrs"]
 
         if max(t_minus_hours) > max_lookback - 1:
             raise ValueError(