+++ src/teehr/models/loading.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class ChunkByEnum(str, Enum):
-    day = "day"
-    location_id = "location_id"